Feat: api 엔드포인트 추가, server.js추가, 연결 url설정 필요
