+++ docs/public/page/itinerary.html
@@ -14,7 +14,7 @@
       rel="stylesheet"
     />
     <!-- 커스텀 스타일 -->
-    <link href="../css/styles.css" rel="stylesheet" />
+    <link href="/css/style.css" rel="stylesheet" />
     <!-- Tailwind CSS 및 설정 -->
     <link
       href="https://ai-public.creatie.ai/gen_page/tailwind-custom.css"
